Moved two path utils to file_utils
This refactoring is needed in order to allow circular imports
on upcoming change.

Related: https://github.com/ansible-community/ansible-lint/pull/1134
